We need floating point width and height.
The 2x image may not be an even number of pixels in size, so we need to
preserve the half-pixels when scaling our frame size into logical pixels. Whee!

+++ core/src/main/java/tripleplay/util/Frames.java
@@ -17,10 +17,10 @@ import playn.core.ImageLayer;
 public interface Frames
 {
     /** Returns the width of a logical frame. */
-    int width ();
+    float width ();
 
     /** Returns the height of a logical frame. */
-    int height ();
+    float height ();
 
     /** Returns the number of frames available. */
     int count ();
